update teleportation services
