+++ src/services/teleportation/index.ts
@@ -1,0 +1,11 @@
+import { TeleportationAddress } from './TeleportationAddress'
+import { TeleportationContract } from './TeleportationContract'
+import { IsTeleportationOfAssetSupported } from './IsTeleportationOfAssetSupported'
+import { DepositWithTeleporter } from './DepositWithTeleporter'
+
+export {
+  TeleportationAddress,
+  TeleportationContract,
+  IsTeleportationOfAssetSupported,
+  DepositWithTeleporter,
+}
